Adding packge structure for memory measuring
